requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu

python-dotenv
amqp
loguru
ffmpeg-python
Pillow
numpy
tritonclient[all]
opencv-python
rich